requirements.txt
alembic==1.4.2
click==7.1.2
dnspython==1.16.0
ecdsa==0.13
email-validator
eventlet==0.25.1
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-Migrate
Flask-SocketIO==4.2.1
Flask-SQLAlchemy
Flask-WTF
Flask-Mail
future
gevent
gunicorn
gevent-websocket
greenlet
gensim==3.8.3
itsdangerous
Jinja2
Mako
MarkupSafe==0.23
pycparser
pycryptodome
python-dateutil
python-editor
python-engineio==3.13.2
python-socketio==4.6.0
pytest-flask
retrying
six
SQLAlchemy
Werkzeug
wincertstore
WTForms
zope.event
zope.interface
pandas
numpy